Fix the compilation on 5.2.0~alpha1 and remove the useless [@@inlined] attribute
It seems that this code try to optimize the way we manipulate `Mtime.t` values.
However, OCaml is not able to do cross-module optimisation: [@@inlined] has no
incidence in the assembler generated (tested with 4.14 and 5.0). This patch
remove the overlooked optimisation and be compatible with the next release of
OCaml 5.2.

.